create cal app and fix some bug in todo and fix routing and othter

+++ ToDoApp/src/app/todos/todos.component.ts
@@ -12,17 +12,12 @@ export class TodosComponent implements OnInit {
 
   todoList: any;
   temp: any;
-  editMode: any;
   activeCount = 0;
   completedCount = 0;
 
   constructor(private route: Router, private todoservice: TodosService, private toastr: ToastrService) {
     this.temp = /list/.test(window.location.href) ? 'list' : (/active/.test(window.location.href) ? 'active' : 'completed');
-    this.todoservice.key.next(this.temp);
 
-    this.todoservice.editMode.subscribe(res => {
-      this.editMode = res;
-    })
   }
 
   ngOnInit(): void {
@@ -37,18 +32,22 @@ export class TodosComponent implements OnInit {
 
     }
   }
+  addToDo() {
+    this.route.navigate(['todo/add']);
+  }
   editTodo(todo: any) {
-    this.todoservice.editMode.next(true);
     this.todoservice.todoObj.next(todo);
     this.route.navigate([`todo/edit`, todo.id]);
   }
 
   deleteTodo(index: any) {
-    this.completedCount--;
-    this.activeCount--;
-    this.todoservice.deleteTodo(index);
-    this.todoList = this.todoservice.getTodos();
-    this.toastr.success('To-Do Deleted Successfully!');
+    if (confirm('Do you really want to Delete this?')) {
+      this.completedCount--;
+      this.activeCount--;
+      this.todoservice.deleteTodo(index);
+      this.todoList = this.todoservice.getTodos();
+      this.toastr.success('To-Do Deleted Successfully!');
+    }
   }
 
   checkedTodo(id: any, status: any) {
